Address CodeScene: split Turn.io form tests out, extract message lookup
Low Cohesion: TestTurnIOMessagingConfigForm was appended to
test_messaging_providers.py, a 920-line file already spanning provider
parsing, service media/audio/window behaviour and Twilio file sending.
A form test does not belong there; it moves to its own file alongside
the existing test_obfusticating_form.py.

Complex Conditional: the defensive guard in
is_non_conversational_whatsapp_message was a three-clause or. Extracted
to _first_whatsapp_message, which returns the first message dict or
None, leaving the caller a single None check.
